rename generate_enum_match_method file to match assist name
