+++ crates/ide_assists/src/handlers/generate_enum_is_method.rs
